use common partial head component for helmet
